--- packages/client/src/services/actions/index.ts
@@ -1,1 +1,0 @@
-export * from './user'
