--- packages/client/src/services/hooks/redux-hooks.ts
@@ -1,12 +1,0 @@
-import {
-  TypedUseSelectorHook,
-  useDispatch as dispatchHook,
-  useSelector as selectorHook,
-} from 'react-redux'
-import { AppDispatch, AppThunk, RootState } from '../../types/common'
-
-// Теперь этот хук «знает» структуру хранилища
-export const useSelector: TypedUseSelectorHook<RootState> = selectorHook
-
-// Хук не даст отправить экшен, который ему не знаком
-export const useDispatch = () => dispatchHook<AppDispatch | AppThunk>()
