+++ packages/client/src/services/redux/login/type/const.ts
@@ -1,0 +1,7 @@
+export const USER_SIGNUP_REQUEST = 'USER_SIGNUP_REQUEST' as const;
+export const USER_SIGNUP_SUCCESS = 'USER_SIGNUP_SUCCESS' as const;
+export const USER_SIGNUP_FAILED = 'USER_SIGNUP_FAILED' as const;
+
+export const LOGIN_REQUEST = 'LOGIN_REQUEST' as const
+export const LOGIN_SUCCESS = 'LOGIN_SUCCESS' as const
+export const LOGIN_FAILED = 'LOGIN_FAILED' as const
