+++ packages/client/src/services/redux/user/type/const.ts
@@ -1,0 +1,11 @@
+export const USER_REQUEST = 'USER_REQUEST' as const
+export const USER_REQUEST_SUCCESS = 'USER_REQUEST_SUCCESS' as const
+export const USER_REQUEST_FAILED = 'USER_REQUEST_FAILED' as const
+
+export const USER_UPDATE_PASSWORD_REQUEST = 'USER_UPDATE_PASSWORD_REQUEST' as const
+export const USER_UPDATE_PASSWORD_SUCCESS = 'USER_UPDATE_PASSWORD_SUCCESS' as const
+export const USER_UPDATE_PASSWORD_FAILED = 'USER_UPDATE_PASSWORD_FAILED' as const
+
+export const UPDATE_USER_REQUEST = 'UPDATE_USER_REQUEST' as const
+export const UPDATE_USER_SUCCESS = 'UPDATE_USER_SUCCESS' as const
+export const UPDATE_USER_FAILED = 'UPDATE_USER_FAILED' as const
