+++ packages/client/src/types/common.ts
@@ -1,9 +1,4 @@
-import { TUserActions } from '../services/actions'
-import { ThunkAction } from 'redux-thunk'
-import { Action, ActionCreator, Dispatch } from 'redux'
-import { TWSActions } from '../services/actions/websocket'
-import { store } from '../services/store/store'
-import { TLoginActions } from '../services/redux/login/actions/login'
+
 
 export type TEvents = Record<string, (e: Event) => void>
 
@@ -18,14 +13,3 @@ export type Range<F extends number, T extends number> = Exclude<
   Enumerate<T>,
   Enumerate<F>
 >
-
-
-
-
-export type TAppActions = TUserActions | TWSActions | TLoginActions
-
-export type RootState = ReturnType<typeof store.getState>
-export type AppThunk<TReturn = void> = ActionCreator<
-  ThunkAction<TReturn, Action, RootState, TAppActions>
->
-export type AppDispatch = Dispatch<TAppActions>
